Remove the hosting site indirection for auth
Now, we call the custom-token generator directly with the acess
token from the client. This makes it so that now, any actual
details are sent securely over HTTPS (except for the initial code
itself which we can't do anything about until and unless Reddit
implements PKCE or TLS/MA or OIDC or something fancy like that
on top of their traditional OAuth2 system).